refactor Map comp. to import styling
