+++ react-front-end/src/components/header.scss
@@ -31,10 +31,5 @@ nav {
     align-self: center;
     min-width: 100px;
     max-width: 120px;
-    // font-family: 'Poppins', sans-serif;
-    // font-family: 'Zilla Slab Highlight', cursive;
-    // color: $white;
-    // font-size: 1.3rem;
-    // opacity: .85;
   }
 }
